+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# __Version 3:__
+1. Bot Will Stop (It will Raise KeyboardInterrupt) if the OwO Bot Messages You. You Won't Be Banned Because Of Some Dumb Captcha YAY!
+
+</br>
+
 # __Version 2:__
 1. Discord RPC Added
 1. Fixed Unseens' Code